+++ sora-android-sdk/src/main/kotlin/jp/shiguredo/sora/sdk/channel/option/SoraMediaOption.kt
@@ -128,8 +128,6 @@ class SoraMediaOption {
     @JvmOverloads
     fun enableSpotlight(option: SoraSpotlightOption, enableSimulcast: Boolean = true) {
         spotlightOption = option
-        multistreamEnabled = true
-
         if (enableSimulcast) {
             enableSimulcast()
         }
@@ -212,25 +210,12 @@ class SoraMediaOption {
     internal val upstreamIsRequired: Boolean
         get() = audioUpstreamEnabled || videoUpstreamEnabled
 
-    internal var _multistreamIsRequired: Boolean? = null
-
-    internal var multistreamIsRequired: Boolean
-        get() = when {
-            _multistreamIsRequired != null ->
-                _multistreamIsRequired!!
-            downstreamIsRequired && upstreamIsRequired ->
-                // 双方向通信の場合は multistream フラグを立てる
-                true
-            else ->
-                // 未設定の場合は true として扱う
-                multistreamEnabled ?: true
-        }
-        set(value) {
-            _multistreamIsRequired = value
-        }
-
+    // TODO(zztkm): internal かつ 未使用なので削除して良い
     internal var _requiredRole: SoraChannelRole? = null
 
+    /**
+     * Upstream と Downstream の設定から、必要なロールを決定します.
+     */
     internal val requiredRole: SoraChannelRole
         get() = if (upstreamIsRequired && downstreamIsRequired)
             SoraChannelRole.SENDRECV